@@ -3,14 +3,15 @@
   "description": "Yii2 INN (UA) Validator",
   "type": "library",
   "require": {
-    "php": ">=7.4",
+    "php": "^8.4",
     "yiisoft/yii2": "^2.0.31",
     "horat1us/inn-parser": "^1.2",
-    "horat1us/yii2-model-schema": "^1.0.1"
+    "horat1us/yii2-model-schema": "^2.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^9.2",
-    "squizlabs/php_codesniffer": "^3.5"
+    "phpunit/phpunit": "^11.0",
+    "squizlabs/php_codesniffer": "^3.10",
+    "horat1us/yii2-asset-free": "^1.0"
   },
   "license": "MIT",
   "authors": [
@@ -34,9 +35,14 @@
     ]
   },
   "scripts": {
-    "lint": "phpcs --standard=PSR2 ./src ./tests",
-    "phpcbf": "phpcbf --standard=PSR2 ./src ./tests",
+    "lint": "phpcs --standard=PSR12 ./src ./tests",
+    "phpcbf": "phpcbf --standard=PSR12 ./src ./tests",
     "cover": "phpunit --coverage-clover=coverage.xml",
     "test": "phpunit"
+  },
+  "config": {
+    "allow-plugins": {
+      "yiisoft/yii2-composer": false
+    }
   }
 }
